Upload Files
Loaded Folders that contain the files of the site.  There are currently three different folders.  One is of the OG site and I can probably get rid of that at some point.  I am just keeping it as a reference incase I really mess something up.  The other is a testing folder to test ideas.  The main folder is called CurrentSite and it is actual files that make the website