Updated Home Page
Updated the home page to have a banner_content width of 750px.  I also updated the banner_content.img to have a width and height of 100% all the way up to the max @media size (which I changed to 2400px). This keeps the photo from blowing up.  And the window size can now only hit 2400px (I think, do not know enough about css to know if that's true).