Slide Show arrows and Dots are below the picture
Using the CSS transform functionality I was able to use transform: translate to place the dots and the arrows into the location I desired.  I then had to add margins to the pictures so that the header and paragraph did not mush in with the arrows when the website is in mobile device layout.